Add event listeners to load saved game state or delete it

+++ main.js
@@ -64,6 +64,7 @@ function renderGame() {
 
     winsP1.innerHTML = `Wins: ${playerOne.wins}`;
     winsP2.innerHTML = `Wins: ${playerTwo.wins}`;
+    loadGame.classList.add("hidden");
 
     for(let i = 0; i < gameBoard.length; i++) {
         gameBoardDOM.children[i].innerHTML = gameBoard[i];
@@ -79,6 +80,7 @@ function startGame() {
         gameBoardDOM.appendChild(cell);
         saveButton.classList.remove("hidden");
         quitButton.classList.remove("hidden");
+        loadGame.classList.add("hidden");
     }
 }
 
@@ -174,6 +176,8 @@ let winsP2 = document.querySelector("#p2Wins");
 
 // Game Board Query Selectors
 let gameBoardDOM = document.querySelector(".game-board");
+let loadGame = document.querySelector(".load-game");
+let savedGame = document.querySelector("#savedGame");
 let saveButton = document.querySelector("#saveButton");
 let quitButton = document.querySelector("#quitButton");
 
@@ -279,23 +283,84 @@ gameBoardDOM.addEventListener("click", (event) => {
     }
 })
 
+// Extension Event Listeners
+
 quitButton.addEventListener("click", () => {
     location.reload();
 })
 
 saveButton.addEventListener("click", () => {
-    let gameState = {
-        p1: playerOne,
-        p2: playerTwo,
-        turnCount: turnCounter,
-        game: gameBoard, 
-        date: Date.now()
+    if(JSON.stringify(localStorage).includes("game")) {
+        let verifyOverwrite = prompt("This will overwrite your previous saved game. Type Y to continue or N to cancel.");
+        if(verifyOverwrite === "Y") {
+            let gameState = {
+                p1: playerOne,
+                p2: playerTwo,
+                turnCount: turnCounter,
+                game: gameBoard, 
+                date: Date.now()
+            }
+        
+            let gameStateSerialized = JSON.stringify(gameState);
+            localStorage.setItem("gameState", gameStateSerialized);
+            let gameStateDeserialized = JSON.parse(localStorage.getItem("gameState"));
+            return gameStateDeserialized;
+        } else if(verifyOverwrite === "N") {
+            return;
+        } else {
+            alert("Invalid response!");
+        }
+    } else {
+        let gameState = {
+            p1: playerOne,
+            p2: playerTwo,
+            turnCount: turnCounter,
+            game: gameBoard, 
+            date: Date.now()
+        }
+    
+        let gameStateSerialized = JSON.stringify(gameState);
+        localStorage.setItem("gameState", gameStateSerialized);
+        let gameStateDeserialized = JSON.parse(localStorage.getItem("gameState"));
+        return gameStateDeserialized;
     }
+})
 
-    let gameStateSerialized = JSON.stringify(gameState);
-    localStorage.setItem("gameState", gameStateSerialized);
-    let gameStateDeserialized = JSON.parse(localStorage.getItem("gameState"));
-    console.log(gameStateDeserialized);
-    return gameStateDeserialized
+addEventListener("load", () => {
+    if(JSON.stringify(localStorage).includes("game")) {
+        loadGame.classList.remove("hidden");
+        let game = document.createElement("div");
+        game.setAttribute("id", "savedGame");
+        game.innerHTML = `Saved Game: ${Date(JSON.parse(localStorage.gameState).date)} <button id="deleteSavedGame">X</button>`;
+        loadGame.appendChild(game);
+    } else {
+        loadGame.classList.add("hidden");
+    }
+})
+
+loadGame.addEventListener("click", (event) => {
+    if(event.target.id === "savedGame") {
+        startGame();
+        gameBoard = JSON.parse(localStorage.gameState).game;
+        turnCounter = JSON.parse(localStorage.gameState).turnCount;
+        playerOne = JSON.parse(localStorage.gameState).p1;
+        playerTwo = JSON.parse(localStorage.gameState).p2;
+        playerTokenP1.innerText = playerOne.token;
+        playerTokenP2.innerText = playerTwo.token;
+            
+        inputSectionP1.classList.add("hidden");
+        inputSectionP2.classList.add("hidden");
+
+        infoSectionP1.classList.remove("hidden");
+        infoSectionP2.classList.remove("hidden");
+        renderGame(); 
+    }
+})
+
+loadGame.addEventListener("click", (event) => {
+    if(event.target.id === "deleteSavedGame") {
+        localStorage.removeItem("gameState");
+        location.reload();
+    }
 })
